+++ bliss/flint/widgets/one_dim_data_plot.py
@@ -4,5 +4,11 @@
 #
 # Copyright (c) 2015-2020 Beamline Control Unit, ESRF
 # Distributed under the GNU LGPLv3. See LICENSE for more info.
-"""Package containing tricky patches of libraries
-"""
+
+from __future__ import annotations
+
+from .mca_plot import McaPlotWidget
+
+
+class OneDimDataPlotWidget(McaPlotWidget):
+    pass