+++ tests/shell/test_standard.py
@@ -5,12 +5,14 @@ import pytest
 import numpy
 import psutil
 
+from bliss.shell import standard
 from bliss.shell.standard import wa, wm, sta, stm, _launch_silx, umv
 
 from bliss.shell.standard import sin, cos, tan, arcsin, arccos, arctan, arctan2
 from bliss.shell.standard import log, log10, sqrt, exp, power, deg2rad, rad2deg
 from bliss.shell.standard import rand, date, sleep
 from bliss.shell.standard import flint
+from bliss.controllers.lima import roi as lima_rois
 
 
 @pytest.fixture
@@ -316,3 +318,26 @@ def test_open_kill_flint(test_session_without_flint):
     process = psutil.Process(pid)
     psutil.wait_procs([process], timeout=1)
     assert not psutil.pid_exists(pid)
+
+
+def test_edit_roi_counters(
+    mocker, beacon, default_session, lima_simulator, test_session_with_flint
+):
+    class PlotMock:
+        def select_shapes(self, *args, **kwargs):
+            roi1 = lima_rois.Roi(10, 11, 100, 101, name="roi1")
+            roi2 = lima_rois.RoiProfile(20, 21, 200, 201, name="roi2", mode="vertical")
+            return [roi1, roi2]
+
+    # Mock few functions to coverage the code without flint
+    mocker.patch("bliss.common.plot.plot_image", return_value=PlotMock())
+
+    cam = beacon.get("lima_simulator")
+
+    cam.roi_counters.clear()
+    cam.roi_profiles.clear()
+    cam.roi_counters["foo1"] = 20, 20, 18, 20
+    cam.roi_profiles["foo2"] = 20, 20, 18, 20, "vertical"
+    standard.edit_roi_counters(cam)
+    assert "roi1" in cam.roi_counters
+    assert "roi2" in cam.roi_profiles